chore: form molecules types moved to globals
