+++ src/models/types/forms/form-checkbox.ts
@@ -1,0 +1,4 @@
+export type FormCheckboxProps = {
+  label: string;
+  name?: string;
+};
